Update all application

+++ app/src/main/java/com/example/futureadmin/AddProductComputerCPUActivity.kt
@@ -78,7 +78,7 @@ class AddProductComputerIntelCPUActivity : AppCompatActivity() {
                     databaseReference.setValue(ComputerCPU(productNumber,
                     name,
                         category,
-                        price,
+                        price.toDouble(),
                         quantity,
                         imageLink,
                         description,
