+++ app/src/main/java/com/example/futureadmin/AddProductComputerRamActivity.kt
@@ -107,7 +107,7 @@ class AddProductComputerRamActivity : AppCompatActivity() {
                         productNumber,
                         name,
                         category,
-                        price,
+                        price.toDouble(),
                         quantity,
                         imageLink,
                         description,
